update .h all malloc secure in ast
