+++ src/exec/redirection.c
@@ -78,3 +78,18 @@ int	apply_all_redirections(t_ast *node)
 	
 	return (0);
 }
+
+void	cleanup_heredoc_files(t_ast *node)
+{
+	if (!node)
+		return ;
+	if (node->token == REDIR_HEREDOC && node->cmd && node->cmd->args)
+	{
+		if (node->cmd->args[0])
+			unlink(node->cmd->args[0]);
+	}
+	if (node->left)
+		cleanup_heredoc_files(node->left);
+	if (node->right)
+		cleanup_heredoc_files(node->right);
+}
